Refactored configuration to remove endpoint property injected values.

+++ components/camel-sjms/src/main/java/org/apache/camel/component/sjms/SjmsProducer.java
@@ -202,7 +202,7 @@ public abstract class SjmsProducer extends DefaultAsyncProducer  {
      * @return the acknowledgment mode
      */
     public int getAcknowledgeMode() {
-        return getSjmsEndpoint().getAcknowledgementMode();
+        return getSjmsEndpoint().getAcknowledgementMode().intValue();
     }
 
     /**
